Reformat and restructure (according FDP)
