+++ src/main/java/solutions/fairdata/fdp/index/api/controller/EntriesController.java
@@ -29,8 +29,8 @@ import org.springframework.data.domain.Pageable;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
-import solutions.fairdata.fdp.index.api.dto.EntryDTO;
-import solutions.fairdata.fdp.index.service.IndexService;
+import solutions.fairdata.fdp.index.api.dto.IndexEntryDTO;
+import solutions.fairdata.fdp.index.service.IndexEntryService;
 
 import java.util.List;
 import java.util.stream.Collectors;
@@ -42,15 +42,15 @@ import java.util.stream.StreamSupport;
 public class EntriesController {
 
     @Autowired
-    private IndexService service;
+    private IndexEntryService service;
 
     @GetMapping("")
-    public Page<EntryDTO> getEntriesPage(Pageable pageable) {
+    public Page<IndexEntryDTO> getEntriesPage(Pageable pageable) {
         return service.getEntriesPage(pageable).map(service::toDTO);
     }
 
     @GetMapping("/all")
-    public List<EntryDTO> getEntriesAll() {
+    public List<IndexEntryDTO> getEntriesAll() {
         return StreamSupport.stream(service.getAllEntries().spliterator(), true).map(service::toDTO).collect(Collectors.toList());
     }
 }
