+++ src/main/java/solutions/fairdata/fdp/index/database/changelogs/DatabaseChangeLog.java
@@ -20,6 +20,16 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
  * THE SOFTWARE.
  */
+package solutions.fairdata.fdp.index.database.changelogs;
 
-@javax.annotation.ParametersAreNonnullByDefault
-package solutions.fairdata.fdp.index.api;
+import com.github.mongobee.changeset.ChangeLog;
+import com.github.mongobee.changeset.ChangeSet;
+import com.mongodb.client.MongoDatabase;
+
+@ChangeLog
+public class DatabaseChangeLog {
+    @ChangeSet(order = "000", id = "initMongoDB", author = "MarekSuchanek")
+    public void initMongoDB(MongoDatabase db) {
+        // Nothing to DO, just "first" making the version
+    }
+}
