+++ src/main/java/solutions/fairdata/fdp/index/database/changelogs/package-info.java
@@ -20,15 +20,6 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
  * THE SOFTWARE.
  */
-package solutions.fairdata.fdp.index.api;
 
-import org.springframework.context.annotation.ComponentScan;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.context.annotation.Import;
-import solutions.fairdata.fdp.index.service.ServiceConfig;
-
-@Configuration
-@ComponentScan
-@Import(ServiceConfig.class)
-public class ApiConfig {
-}
+@javax.annotation.ParametersAreNonnullByDefault
+package solutions.fairdata.fdp.index.database.changelogs;
