+++ src/main/java/solutions/fairdata/fdp/index/entity/IndexEntry.java
@@ -20,6 +20,19 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
  * THE SOFTWARE.
  */
+package solutions.fairdata.fdp.index.entity;
 
-@javax.annotation.ParametersAreNonnullByDefault
-package solutions.fairdata.fdp.index.domain;
+import lombok.Data;
+import org.bson.types.ObjectId;
+import org.springframework.data.annotation.Id;
+import org.springframework.data.mongodb.core.mapping.Document;
+
+@Document
+@Data
+public class IndexEntry {
+    @Id
+    protected ObjectId id;
+    private String clientUrl;
+    private String registrationTime;
+    private String modificationTime;
+}
