+++ src/main/java/solutions/fairdata/fdp/index/package-info.java
@@ -20,24 +20,6 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
  * THE SOFTWARE.
  */
-package solutions.fairdata.fdp.index.api.dto;
 
-import io.swagger.v3.oas.annotations.media.Schema;
-import lombok.Data;
-import org.hibernate.validator.constraints.URL;
-import javax.validation.constraints.NotNull;
-import java.time.OffsetDateTime;
-
-@Data
-@Schema(name = "Entry")
-public class EntryDTO {
-    @NotNull
-    @URL
-    private String clientUrl;
-
-    @NotNull
-    private OffsetDateTime registrationTime;
-
-    @NotNull
-    private OffsetDateTime modificationTime;
-}
+@javax.annotation.ParametersAreNonnullByDefault
+package solutions.fairdata.fdp.index;
