+++ src/main/resources/static/js/localize-timestamps.js
@@ -1,5 +1,5 @@
 jQuery(document).ready(() => {
-    jQuery("td.timestamp").each(function() {
+    jQuery("td.timestamp").each(function () {
         const dt = new Date($(this).text());
         $(this).text(dt.toLocaleString("en-GB"));
     });
