+++ src/main/resources/templates/fragments/header.html
@@ -4,7 +4,7 @@
 <body>
 <header>
     <div class="container">
-        <img src="../static/img/fair_logo.png" th:src="@{/img/fair_logo.png}" alt="logo" class="logo" />
+        <img alt="logo" class="logo" src="../static/img/fair_logo.png" th:src="@{/img/fair_logo.png}"/>
         <h1>FAIR Data Point index</h1>
     </div>
     <div class="separator"></div>
