+++ src/main/scss/site.scss
@@ -54,6 +54,7 @@ footer div.container {
 table.entries td.endpoint {
   overflow-wrap: anywhere;
 }
+
 table.entries td.timestamp,
 table.entries th.right {
   text-align: right;
